+++ geulkkoli/src/main/java/com/geulkkoli/domain/user/service/LoginService.java
@@ -1,6 +1,6 @@
-package com.geulkkoli.domain.service;
+package com.geulkkoli.domain.user.service;
 
-import com.geulkkoli.domain.User;
+import com.geulkkoli.domain.user.User;
 import com.geulkkoli.infrastructure.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
