+++ geulkkoli/src/main/java/com/geulkkoli/infrastructure/ImplUserRepository.java
@@ -1,6 +1,6 @@
 package com.geulkkoli.infrastructure;
 
-import com.geulkkoli.domain.User;
+import com.geulkkoli.domain.user.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
