+++ geulkkoli/src/test/java/com/geulkkoli/domain/PostTest.java
@@ -1,5 +1,6 @@
 package com.geulkkoli.domain;
 
+import com.geulkkoli.domain.post.Post;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
