+++ geulkkoli/src/test/java/com/geulkkoli/domain/UserTest.java
@@ -1,5 +1,6 @@
 package com.geulkkoli.domain;
 
+import com.geulkkoli.domain.user.User;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
