+++ geulkkoli/src/test/java/com/geulkkoli/infrastructure/PostRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.geulkkoli.infrastructure;
 
-import com.geulkkoli.domain.Post;
+import com.geulkkoli.domain.post.Post;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
